Bug: 139893257

+++ tests/wifitests/src/com/android/server/wifi/WifiConnectivityHelperTest.java
@@ -22,7 +22,8 @@ import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
 
 import android.app.test.MockAnswerUtil.AnswerWithArguments;
-import android.support.test.filters.SmallTest;
+
+import androidx.test.filters.SmallTest;
 
 import org.junit.After;
 import org.junit.Before;
@@ -64,7 +65,8 @@ public class WifiConnectivityHelperTest {
 
     private void setupWifiNative() {
         // Return firmware roaming feature as supported by default.
-        when(mWifiNative.getSupportedFeatureSet(any())).thenReturn(WIFI_FEATURE_CONTROL_ROAMING);
+        when(mWifiNative.getSupportedFeatureSet(any()))
+                .thenReturn((long) WIFI_FEATURE_CONTROL_ROAMING);
 
         doAnswer(new AnswerWithArguments() {
             public boolean answer(String ifaceName, WifiNative.RoamingCapabilities roamCap)
@@ -118,7 +120,8 @@ public class WifiConnectivityHelperTest {
      */
     @Test
     public void returnFirmwareRoamingNotSupported() {
-        when(mWifiNative.getSupportedFeatureSet(any())).thenReturn(~WIFI_FEATURE_CONTROL_ROAMING);
+        when(mWifiNative.getSupportedFeatureSet(any()))
+                .thenReturn((long) ~WIFI_FEATURE_CONTROL_ROAMING);
         assertTrue(mWifiConnectivityHelper.getFirmwareRoamingInfo());
         assertFalse(mWifiConnectivityHelper.isFirmwareRoamingSupported());
     }
